Add Store.Versions()
This can be used by compacter and other code wanted to get a list of all available state versions
